backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# AWS & AI
boto3==1.34.0
# strands-agents  # Uncomment when available

# Code Parsing
javalang==0.13.0
tree-sitter==0.20.4
GitPython==3.1.40

# Async & Threading
asyncio==3.4.3
aiofiles==23.2.1

# Utils
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
Jinja2==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
pylint==3.0.3
